audio-kernel: Implement Xiaomi audio hardware registration
The Xiaomi mmhardware device driver is responsible for registering the
specific audio hardware present on the Xiaomi device within the
MIUI audio framework.

For instance, if the device supports adsp hardware,
it will be registered through this driver, allowing vendor binaries to read it
for further processing.

Changes made to the vendor driver include:
- Removal of Android.mk
- Cleanup of mmhardware_sysfs error logging
- Utilization of snprintf instead of sprintf
- Semantic improvements in mm_register_show

Change-Id: I9448bfb87ca968af732b0fc4b2d005a7e3501bcb
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>